requirements.txt
Flask
pymongo
yagmail
requests
apscheduler
python-dotenv
google-genai  
pytz  
gunicorn